fix | moving authOptions into separated auth file and exporting from there
