+++ app/actions/get-session.ts
@@ -1,8 +1,8 @@
 import { getServerSession } from 'next-auth';
+import { authOptions } from '../../auth'
 
-import { authOptions } from '../api/auth/[...nextauth]/route';
-
+// import { authOptions } from '../api/auth/[...nextauth]/route';
 
 export default async function getSession() {
-  return await getServerSession(authOptions)
+  return await getServerSession(authOptions);
 }
