+++ app/api/auth/[...nextauth]/route.ts
@@ -5,59 +5,61 @@ import GithubProvider from 'next-auth/providers/github';
 import GoogleProvider from 'next-auth/providers/google';
 import { PrismaAdapter } from '@next-auth/prisma-adapter';
 
+import { authOptions } from '../../../../auth';
+
 import prisma from '@/app/libs/prismadb';
 
-export const authOptions: NextAuthOptions = {
-  adapter: PrismaAdapter(prisma),
-  providers: [
-    GithubProvider({
-      clientId: process.env.GITHUB_ID as string,
-      clientSecret: process.env.GITHUB_SECRET as string,
-    }),
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID as string,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
-    }),
-    CredentialsProvider({
-      name: 'credentials',
-      credentials: {
-        email: { label: 'email', type: 'text' },
-        password: { label: 'password', type: 'password' },
-      },
-      async authorize(credentials) {
-        if (!credentials?.email || !credentials?.password) {
-          throw new Error('Invalid Credentials');
-        }
+// export const authOptions: NextAuthOptions = {
+//   adapter: PrismaAdapter(prisma),
+//   providers: [
+//     GithubProvider({
+//       clientId: process.env.GITHUB_ID as string,
+//       clientSecret: process.env.GITHUB_SECRET as string,
+//     }),
+//     GoogleProvider({
+//       clientId: process.env.GOOGLE_CLIENT_ID as string,
+//       clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
+//     }),
+//     CredentialsProvider({
+//       name: 'credentials',
+//       credentials: {
+//         email: { label: 'email', type: 'text' },
+//         password: { label: 'password', type: 'password' },
+//       },
+//       async authorize(credentials) {
+//         if (!credentials?.email || !credentials?.password) {
+//           throw new Error('Invalid Credentials');
+//         }
 
-        const user = await prisma.user.findUnique({
-          where: {
-            email: credentials.email,
-          },
-        });
+//         const user = await prisma.user.findUnique({
+//           where: {
+//             email: credentials.email,
+//           },
+//         });
 
-        if (!user || !user?.hashedPassword) {
-          throw new Error('Invalid Credentials');
-        }
+//         if (!user || !user?.hashedPassword) {
+//           throw new Error('Invalid Credentials');
+//         }
 
-        const isCorrectPassword = await bcrypt.compare(
-          credentials.password,
-          user.hashedPassword
-        );
+//         const isCorrectPassword = await bcrypt.compare(
+//           credentials.password,
+//           user.hashedPassword
+//         );
 
-        if (!isCorrectPassword) {
-          throw new Error('Invalid Credentials');
-        }
+//         if (!isCorrectPassword) {
+//           throw new Error('Invalid Credentials');
+//         }
 
-        return user;
-      },
-    }),
-  ],
-  debug: process.env.NODE_ENV === 'development',
-  session: {
-    strategy: 'jwt',
-  },
-  secret: process.env.NEXTAUTH_SECRET,
-};
+//         return user;
+//       },
+//     }),
+//   ],
+//   debug: process.env.NODE_ENV === 'development',
+//   session: {
+//     strategy: 'jwt',
+//   },
+//   secret: process.env.NEXTAUTH_SECRET,
+// };
 
 const handler = NextAuth(authOptions);
